refactor: move managers to tools namespace
